+++ src/bot/edax/mod.rs
@@ -1,9 +1,7 @@
 pub mod bot;
 pub mod r#const;
 pub mod eval;
-pub mod r#move;
 pub mod node;
 pub mod search;
 pub mod search_state;
-pub mod square;
 pub mod weights;
